+++ packages/biconomy-middleware/src/types.ts
@@ -14,6 +14,27 @@ export type ContractMetadata = {
 	[key: string]: any
 }
 
+type BiconomyApiCommonResponse = {
+	code: number
+	message: string
+}
+
+export type BiconomyApiLimitResponse = BiconomyApiCommonResponse & ({
+	limit: {
+		allowed: boolean
+		limitLeft: number
+		resetTime: number
+		type: number
+	}
+	allowed: false
+} | {
+	allowed: true
+})
+
 export interface IContractRegistry {
 	getMetadata(address: string, data?: string): Promise<ContractMetadata | undefined>
+}
+
+export interface ILimitsRegistry {
+	checkLimits(userAddress: string): Promise<BiconomyApiLimitResponse>
 }